move to legacy: weave_client
