+++ weave/integrations/cerebras/cerebras_test.py
@@ -21,7 +21,7 @@ def _get_call_output(call: tsi.CallSchema) -> Any:
 @pytest.mark.vcr(
     filter_headers=["authorization"], allowed_hosts=["api.wandb.ai", "localhost"]
 )
-def test_cerebras_sync(client: weave.weave_client.WeaveClient) -> None:
+def test_cerebras_sync(client: weave.trace.weave_client.WeaveClient) -> None:
     api_key = os.environ.get("CEREBRAS_API_KEY", "DUMMY_API_KEY")
     cerebras_client = Cerebras(api_key=api_key)
 
@@ -56,7 +56,7 @@ def test_cerebras_sync(client: weave.weave_client.WeaveClient) -> None:
     filter_headers=["authorization"], allowed_hosts=["api.wandb.ai", "localhost"]
 )
 @pytest.mark.asyncio
-async def test_cerebras_async(client: weave.weave_client.WeaveClient) -> None:
+async def test_cerebras_async(client: weave.trace.weave_client.WeaveClient) -> None:
     api_key = os.environ.get("CEREBRAS_API_KEY", "DUMMY_API_KEY")
     cerebras_client = AsyncCerebras(api_key=api_key)
