+++ weave/integrations/dspy/dspy_test.py
@@ -4,8 +4,8 @@ from typing import Any, Optional
 import pytest
 
 import weave
+from weave.trace.weave_client import WeaveClient
 from weave.trace_server import trace_server_interface as tsi
-from weave.weave_client import WeaveClient
 
 
 def _get_call_output(call: tsi.CallSchema) -> Any:
