+++ weave/integrations/llamaindex/llamaindex.py
@@ -1,6 +1,6 @@
 from weave.client_context import weave_client as weave_client_context
 from weave.trace.patcher import Patcher
-from weave.weave_client import Call
+from weave.trace.weave_client import Call
 
 TRANSFORM_EMBEDDINGS = False
 ALLOWED_ROOT_EVENT_TYPES = ("query",)
