+++ weave/integrations/mistral/mistral_test.py
@@ -24,7 +24,7 @@ def _get_call_output(call: tsi.CallSchema) -> Any:
 @pytest.mark.vcr(
     filter_headers=["authorization"], allowed_hosts=["api.wandb.ai", "localhost"]
 )
-def test_mistral_quickstart(client: weave.weave_client.WeaveClient) -> None:
+def test_mistral_quickstart(client: weave.trace.weave_client.WeaveClient) -> None:
     # This is taken directly from https://docs.mistral.ai/getting-started/quickstart/
     api_key = os.environ.get("MISTRAL_API_KEY", "DUMMY_API_KEY")
     model = "mistral-large-latest"
@@ -77,7 +77,9 @@ def test_mistral_quickstart(client: weave.weave_client.WeaveClient) -> None:
     filter_headers=["authorization"], allowed_hosts=["api.wandb.ai", "localhost"]
 )
 @pytest.mark.asyncio
-async def test_mistral_quickstart_async(client: weave.weave_client.WeaveClient) -> None:
+async def test_mistral_quickstart_async(
+    client: weave.trace.weave_client.WeaveClient,
+) -> None:
     # This is taken directly from https://docs.mistral.ai/getting-started/quickstart/
     api_key = os.environ.get("MISTRAL_API_KEY", "DUMMY_API_KEY")
     model = "mistral-large-latest"
@@ -125,7 +127,9 @@ Ultimately, the best French cheese is a matter of personal taste. I would recomm
 @pytest.mark.vcr(
     filter_headers=["authorization"], allowed_hosts=["api.wandb.ai", "localhost"]
 )
-def test_mistral_quickstart_with_stream(client: weave.weave_client.WeaveClient) -> None:
+def test_mistral_quickstart_with_stream(
+    client: weave.trace.weave_client.WeaveClient,
+) -> None:
     # This is taken directly from https://docs.mistral.ai/getting-started/quickstart/
     api_key = os.environ.get("MISTRAL_API_KEY", "DUMMY_API_KEY")
     model = "mistral-large-latest"
@@ -182,7 +186,7 @@ def test_mistral_quickstart_with_stream(client: weave.weave_client.WeaveClient) 
 )
 @pytest.mark.asyncio
 async def test_mistral_quickstart_with_stream_async(
-    client: weave.weave_client.WeaveClient,
+    client: weave.trace.weave_client.WeaveClient,
 ) -> None:
     # This is taken directly from https://docs.mistral.ai/getting-started/quickstart/
     api_key = os.environ.get("MISTRAL_API_KEY", "DUMMY_API_KEY")
