+++ weave/ref_base.py
@@ -17,8 +17,8 @@ from .legacy import uris
 REFS: weakref.WeakValueDictionary[int, "Ref"] = weakref.WeakValueDictionary()
 
 if typing.TYPE_CHECKING:
-    from . import weave_client
     from . import weave_types as types
+    from .trace import weave_client
 
 
 def _map_to_ref_strs(obj: typing.Any) -> typing.Any:
