+++ weave/tests/trace/test_op_call_method.py
@@ -13,7 +13,7 @@ def test_op_call_method(client):
 
     # call that returns a Call obj
     res2, call = add.call(1, 2)
-    assert isinstance(call, weave.weave_client.Call)
+    assert isinstance(call, weave.trace.weave_client.Call)
     assert call.inputs == {"a": 1, "b": 2}
     assert call.output == 3
     assert res2 == 3
