+++ weave/tests/trace/test_weave_client.py
@@ -11,9 +11,9 @@ import requests
 
 import weave
 import weave.trace_server.trace_server_interface as tsi
-from weave import Evaluation, weave_client
+from weave import Evaluation
 from weave.legacy import op_def
-from weave.trace import refs
+from weave.trace import refs, weave_client
 from weave.trace.isinstance import weave_isinstance
 from weave.trace.op import Op
 from weave.trace.refs import (
