+++ weave/trace/tests/test_op_return_forms.py
@@ -2,7 +2,7 @@ import pytest
 
 import weave
 from weave.trace.op_extensions.accumulator import add_accumulator
-from weave.weave_client import get_ref
+from weave.trace.weave_client import get_ref
 
 from ...trace_server import trace_server_interface as tsi
