+++ weave/type_serializers/Image/image_test.py
@@ -1,7 +1,7 @@
 from PIL import Image
 
 import weave
-from weave.weave_client import WeaveClient, get_ref
+from weave.trace.weave_client import WeaveClient, get_ref
 
 """When testing types, it is important to test:
 Objects:
